+++ src/configuration/models/unite/uniteDirectories.ts
@@ -6,6 +6,7 @@ import { UniteWwwDirectories } from "./uniteWwwDirectories";
 export class UniteDirectories {
     public wwwRoot: string;
     public packagedRoot: string;
+    public platformRoot: string;
 
     public www: UniteWwwDirectories;
 }